Module Tezos_020_PsParisC_test_helpers

module Account : sig ... end

module Account_helpers : sig ... end

Account_helpers defines a type abstracting the information of an account in the protocol. This includes its pkh, delegate, any funds, staking parameters, etc...

module Adaptive_issuance_helpers : sig ... end

module Assert : sig ... end

module Big_map_helpers : sig ... end

module Block : sig ... end

module Consensus_helpers : sig ... end

module Constants_helpers : sig ... end

module Context : sig ... end

module Contract_helpers : sig ... end

module Cpmm_logic : sig ... end

module Cpmm_repr : sig ... end

module Dal_helpers : sig ... end

module Dummy_zk_rollup : sig ... end

module Error_helpers : sig ... end

Helpers to check expected errors.

module Expr : sig ... end

module Expr_common : sig ... end

module Incremental : sig ... end

module Liquidity_baking_generator : sig ... end

This module provides a set of abstractions to reason about the so-called “liquidity baking” feature1.

module Liquidity_baking_machine : sig ... end

This module provides the means to test extensively the Liquidity Baking (LB) feature. We recall that this feature is built upon three smart contracts: (1) a CPMM contract initially based on Dexter 2, and (2) two tokens contracts. Our objective is to run “scenarios” consisting in interleaved, realistic calls to these contracts, and to assert these scenarios do not yield any undesirable behaviors.

module Log_helpers : sig ... end

module Lqt_fa12_repr : sig ... end

module Lwt_result_wrap_syntax : sig ... end

Extends the Lwt_result_syntax with additional utilities for wrapping results produced by the protocol, i.e. Environment.Error_monad.tzresult, to tzresult Lwt.t values used in the tests.

module Merkle_list_helper : sig ... end

module Nonce : sig ... end

module Op : sig ... end

module Operation_generator : sig ... end

module Result_wrap_syntax : sig ... end

Extends the Result_syntax with additional utilities for wrapping results produced by the protocol, i.e. Environment.Error_monad.tzresult, to tzresult values used in the tests.

module Rewards : sig ... end

The tables are precomputed using this the following formulas:

module Sapling_helpers : sig ... end

module Sc_rollup_helpers : sig ... end

module Scenario : sig ... end

This module centralises all modules related to Scenario writing and execution. Most scenario tests would use most if not all of them, so they only need to open Scenario.

module Scenario_bake : sig ... end

module Scenario_base : sig ... end

This module gathers the basic operations used in test scenarios. This includes starting a scenario, baking, checking and manipulating the state, and various wait functions

module Scenario_begin : sig ... end

module Scenario_constants : sig ... end

module Scenario_dsl : sig ... end

module Scenario_op : sig ... end

This module gathers many protocol operations in the form of scenarios. This includes (but is not limited to) transfers and such (stake, unstake...), as well as various ways to forge double signings.

module Script_big_map : sig ... end

module Script_map : sig ... end

module Script_set : sig ... end

module Slashing_helpers : sig ... end

Helpers related to denunciations and slashing.

module State : sig ... end

module State_account : sig ... end

State_account is dedicated to operations in State.t that would modify an account_state. This includes any operation like transfer, stake, unstake, set_delegate_parameters, anything related to slashing or rewards, etc...

module State_ai_flags : sig ... end

This module handles the logic of flags for AI/AS in the State

module Test_global_constants : sig ... end

module Test_tez : sig ... end

module Testable : sig ... end

module Tez_helpers : sig ... end

module Tez_staking_helpers : sig ... end

Tez_staking_helpers defines different kinds of tez Modules that manipulate them in different ways. They involve more complicated operations, as they are related to staking, thus are represented as partial amounts, and are related to pseudotokens.

module Tezt_helpers : sig ... end

module Ticket_helpers : sig ... end

module Transfers : sig ... end

module Zk_rollup_l2_helpers : sig ... end