Module Tezos_020_PsParisC_test_helpers.Assert

val error : loc:string -> ('a, Tezos_base.TzPervasives.tztrace) Stdlib.result -> (Tezos_error_monad.TzCore.error -> bool) -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val join_errors : (unit, 'a list) Stdlib.result -> (unit, 'a list) Stdlib.result -> (unit, 'a list) Stdlib.result Lwt.t

val test_error_encodings : Tezos_protocol_environment_020_PsParisC.Error_monad.error -> unit

val proto_error : loc:string -> ('a, Tezos_base.TzPervasives.tztrace) Stdlib.result -> (Tezos_protocol_020_PsParisC.Environment.Error_monad.error -> bool) -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val proto_error_with_info : ?error_info_field:[< `Description | `Id | `Message | `Title Title ] -> loc:string -> ('a, Tezos_base.TzPervasives.tztrace) Stdlib.result -> Tezos_base.TzPervasives.String.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val as_proto_error : Tezos_base.TzPervasives.error -> (Tezos_protocol_020_PsParisC.Environment.Error_monad.error, Tezos_base.TzPervasives.error) Stdlib.result

val expect_error : loc:string -> Tezos_base.TzPervasives.Error_monad.tztrace -> (Tezos_protocol_020_PsParisC.Environment.Error_monad.error list -> bool) -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Similar to proto_error, except that errs is directly an error trace instead of a tzresult.

expect_error ~loc errs has the right type to be used as the expect_failure or expect_apply_failure argument of Incremental.add_operation.

val equal : loc:string -> ('a -> 'a -> bool) -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> 'a -> 'a -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val leq : loc:string -> ('a -> 'a -> int) -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> 'a -> 'a -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val lt : loc:string -> ('a -> 'a -> int) -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> 'a -> 'a -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val not_equal : loc:string -> ('a -> 'a -> bool) -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> 'a -> 'a -> unit Tezos_base.TzPervasives.tzresult Lwt.t

module Int32 : sig ... end

module Int64 : sig ... end

val equal_char : loc:string -> Stdlib.Char.t -> Stdlib.Char.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_int : loc:string -> int -> int -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val not_equal_int : loc:string -> int -> int -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val leq_int : loc:string -> int -> int -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_int32 : loc:string -> int32 -> int32 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val leq_int32 : loc:string -> int32 -> int32 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val lt_int32 : loc:string -> int32 -> int32 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_int64 : loc:string -> int64 -> int64 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val not_equal_int64 : loc:string -> int64 -> int64 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val leq_int64 : loc:string -> int64 -> int64 -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_z : loc:string -> Z.t -> Z.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_q : loc:string -> Q.t -> Q.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_bool : loc:string -> bool -> bool -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val not_equal_bool : loc:string -> bool -> bool -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val is_true : loc:string -> bool -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_string : loc:string -> string -> string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val not_equal_string : loc:string -> string -> string -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_tez : loc:string -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val not_equal_tez : loc:string -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_pkh : loc:string -> Tezos_base.TzPervasives.Signature.Public_key_hash.t -> Tezos_base.TzPervasives.Signature.Public_key_hash.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val not_equal_pkh : loc:string -> Tezos_base.TzPervasives.Signature.Public_key_hash.t -> Tezos_base.TzPervasives.Signature.Public_key_hash.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_protocol_hash : loc:string -> Tezos_base.TzPervasives.Protocol_hash.t -> Tezos_base.TzPervasives.Protocol_hash.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val not_equal_protocol_hash : loc:string -> Tezos_base.TzPervasives.Protocol_hash.t -> Tezos_base.TzPervasives.Protocol_hash.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val get_some : loc:string -> 'a option -> 'a Tezos_base.TzPervasives.tzresult Lwt.t

val is_none : loc:string -> pp:(Stdlib.Format.formatter -> 'a -> unit) -> 'a option -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_result : loc:string -> pp_ok:(Stdlib.Format.formatter -> 'a -> unit) -> pp_error:(Stdlib.Format.formatter -> 'b -> unit) -> ('a -> 'a -> bool) -> ('b -> 'b -> bool) -> ('a, 'b) Stdlib.result -> ('a, 'b) Stdlib.result -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val is_error : loc:string -> pp:(Stdlib.Format.formatter -> 'a -> unit) -> ('a, 'b) Stdlib.result -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val get_ok : __LOC__:string -> ('a, Tezos_base.TzPervasives.tztrace) Stdlib.result -> 'a Tezos_base.TzPervasives.tzresult Lwt.t

val contract_property_is : ('a -> 'b -> (Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t) -> loc:string -> 'a -> 'b -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val balance_is : loc:string -> Tezos_020_PsParisC_test_helpers__Context.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

balance_is b c amount checks that the current balance b of contract c is amount.

val frozen_bonds_is : loc:string -> Tezos_020_PsParisC_test_helpers__Context.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

frozen_bonds_is b c amount checks that the current frozen bonds of contract c is amount.

val balance_or_frozen_bonds_was_operated : is_balance:bool -> operand: ('a -> 'b -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t Tezos_protocol_020_PsParisC.Environment.Error_monad.tzresult) -> loc:string -> Tezos_020_PsParisC_test_helpers__Context.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> 'a -> 'b -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val balance_was_credited : loc:string -> Tezos_020_PsParisC_test_helpers__Context.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

balance_was_credited ~loc ctxt contract old_balance amount checks that contract's balance was credited amount tez in comparison to old_balance.

val balance_was_debited : loc:string -> Tezos_020_PsParisC_test_helpers__Context.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

balance_was_credited ~loc ctxt contract old_balance amount checks that contract's balance was debited amount tez in comparison to old_balance.

val frozen_bonds_was_credited : loc:string -> Tezos_020_PsParisC_test_helpers__Context.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

frozen_bonds_was_credited ~loc ctxt contract old_balance amount checks that contract's frozen bonds was credited amount tez in comparison to old_balance.

val frozen_bonds_was_debited : loc:string -> Tezos_020_PsParisC_test_helpers__Context.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> unit Tezos_base.TzPervasives.tzresult Lwt.t

frozen_bonds_was_credited ~loc ctxt contract old_balance amount checks that contract's frozen bonds was credited amount tez in comparison to old_balance.

val pp_print_list : (Stdlib.Format.formatter -> 'a -> unit) -> Stdlib.Format.formatter -> 'a list -> unit

val assert_equal_list : loc:string -> ('a -> 'a -> bool) -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> 'a list -> 'a list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val assert_equal_list_opt : loc:string -> ('a -> 'a -> bool) -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> 'a list option -> 'a list option -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val equal_list_any_order : loc:string -> compare:('a -> 'a -> int) -> string -> (Stdlib.Format.formatter -> 'a -> unit) -> 'a list -> 'a list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Checks that both lists have the same elements, not taking the order of these elements into account, but taking their multiplicity into account.

val to_json_string : 'a Data_encoding.Encoding.t -> 'a -> string

val equal_with_encoding : loc:string -> 'a Data_encoding.Encoding.t -> 'a -> 'a -> unit Tezos_base.TzPervasives.tzresult Lwt.t

val not_equal_with_encoding : loc:string -> 'a Data_encoding.Encoding.t -> 'a -> 'a -> unit Tezos_base.TzPervasives.tzresult Lwt.t