Module Block.Forge

val contents : ?proof_of_work_threshold:Stdlib.Int64.t -> ?seed_nonce_hash:Tezos_protocol_020_PsParisC.Protocol.Nonce_hash.t -> ?liquidity_baking_toggle_vote: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Per_block_votes.per_block_vote -> ?adaptive_issuance_vote: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Per_block_votes.per_block_vote -> payload_hash:Tezos_protocol_020_PsParisC.Protocol.Block_payload_hash.t -> payload_round:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Round.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Block_header.shell_header -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Block_header.contents Tezos_base.TzPervasives.tzresult Lwt.t

type header

val classify_operations : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.packed_operation list -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.packed_operation list list

val forge_header : ?locked_round: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Round.t option -> ?payload_round: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Round.t option -> ?policy:baker_policy -> ?timestamp:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Timestamp.time -> ?operations: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Operation.packed list -> ?liquidity_baking_toggle_vote: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Per_block_votes.per_block_vote -> ?adaptive_issuance_vote: Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Per_block_votes.per_block_vote -> t -> header Tezos_base.TzPervasives.tzresult Lwt.t

Forges a correct header following the policy. The header can then be modified and applied with apply.

val set_seed_nonce_hash : ?proof_of_work_threshold:int64 -> Tezos_protocol_020_PsParisC.Protocol.Nonce_hash.t option -> header -> (header, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

Sets uniquely seed_nonce_hash of a header

val set_baker : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.public_key_hash -> ?consensus_key:Tezos_base.TzPervasives.Signature.public_key_hash -> header -> header

Sets the baker that will sign the header to an arbitrary pkh

val sign_header : header -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Block_header.block_header Tezos_base.TzPervasives.tzresult Lwt.t

Signs the header with the key of the baker configured in the header. The header can no longer be modified, only applied.