Module Tezos_020_PsParisC_test_helpers.Consensus_helpers

type mode =

| Application

| Construction

| Mempool

val show_mode : mode -> string

type kind =

| Preattestation

| Attestation

val test_consensus_operation : ?delegate:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.public_key_hash -> ?slot:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Slot.t -> ?level:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Raw_level.t -> ?round:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Round.t -> ?block_payload_hash:Tezos_protocol_020_PsParisC.Protocol.Block_payload_hash.t -> ?branch:Tezos_base.TzPervasives.Block_hash.t -> attested_block:Block.t -> ?predecessor:Block.t -> ?error:(Tezos_protocol_020_PsParisC.Environment.Error_monad.error -> bool) -> loc:string -> kind -> mode -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

Craft an attestation or preattestation, and bake a block containing it (in application or construction modes) or inject it into a mempool. When error is None, check that it succeeds, otherwise check that it fails as specified by error.

By default, the (pre)attestation is for the first slot and is signed by the delegate that owns this slot. Moreover, the operation points to the given attested_block: in other words, it has that block's level, round, payload hash, and its branch is the predecessor of that block. Optional arguments allow to override these default parameters.

The predecessor is used as the predecessor of the baked block or the head of the mempool. When it is not provided, we use the attested_block for this.

val test_consensus_operation_all_modes_different_outcomes : ?delegate:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.public_key_hash -> ?slot:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Slot.t -> ?level:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Raw_level.t -> ?round:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Round.t -> ?block_payload_hash:Tezos_protocol_020_PsParisC.Protocol.Block_payload_hash.t -> ?branch:Tezos_base.TzPervasives.Block_hash.t -> attested_block:Block.t -> ?predecessor:Block.t -> loc:string -> ?application_error: (Tezos_protocol_020_PsParisC.Environment.Error_monad.error -> bool) -> ?construction_error: (Tezos_protocol_020_PsParisC.Environment.Error_monad.error -> bool) -> ?mempool_error: (Tezos_protocol_020_PsParisC.Environment.Error_monad.error -> bool) -> kind -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val test_consensus_operation_all_modes : ?delegate:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.public_key_hash -> ?slot:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Slot.t -> ?level:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Raw_level.t -> ?round:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Round.t -> ?block_payload_hash:Tezos_protocol_020_PsParisC.Protocol.Block_payload_hash.t -> ?branch:Tezos_base.TzPervasives.Block_hash.t -> attested_block:Block.t -> ?predecessor:Block.t -> ?error:(Tezos_protocol_020_PsParisC.Environment.Error_monad.error -> bool) -> loc:string -> kind -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val delegate_of_first_slot : Context.t -> (Tezos_protocol_020_PsParisC.Environment.Signature.public_key_hash * Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Slot.t, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val delegate_of_slot : ?different_slot:bool -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Slot.t -> Context.t -> (Tezos_protocol_020_PsParisC.Environment.Signature.public_key_hash, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t

val test_consensus_op_for_next : genesis:Block.t -> kind:[< `Attestation | `Preattestation ] -> next:[< `Level | `Round ] -> (unit, Tezos_base.TzPervasives.tztrace) Stdlib.result Lwt.t