Module Context.Contract

val pp : Tezos_protocol_020_PsParisC.Environment.Format.formatter -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> unit

val equal : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> bool

val pkh : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.public_key_hash

val balance : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

Returns the balance of a contract, by default the main balance. If the contract is implicit the frozen balances are available too: deposit, fees or rewards.

val frozen_bonds : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val balance_and_frozen_bonds : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val staked_balance : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t option Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val unstaked_frozen_balance : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t option Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val unstaked_finalizable_balance : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t option Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val full_balance : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val staking_numerator : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Environment.Z.t Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val counter : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Manager_counter.t Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val manager : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Account.t Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val is_manager_key_revealed : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> bool Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val delegate : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.public_key_hash Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val delegate_opt : t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.public_key_hash option Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val storage : t -> Tezos_protocol_020_PsParisC.Protocol.Contract_hash.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Script.expr Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val script : t -> Tezos_protocol_020_PsParisC.Protocol.Contract_hash.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Script.expr Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t

val script_hash : t -> Tezos_protocol_020_PsParisC.Protocol.Contract_hash.t -> Tezos_protocol_020_PsParisC.Protocol.Script_expr_hash.t Tezos_base.TzPervasives.tzresult Tezos_protocol_020_PsParisC.Environment.Lwt.t