Module ValidationMachine.Concrete

A collections of functions to introspect the concrete part of the ValidationMachine state.

val get_xtz_balance : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> t -> xtz Tezos_base.TzPervasives.tzresult Lwt.t

get_xtz_balance c state returns the amount of mutez owned by c in the concrete part of state.

val get_tzbtc_balance : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t env -> t -> tzbtc Tezos_base.TzPervasives.tzresult Lwt.t

get_tzbtc_balance c env state returns the amount of TzBTC owned by c in the concrete part of state, according to the TzBTC contract.

val get_liquidity_balance : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t env -> t -> liquidity Tezos_base.TzPervasives.tzresult Lwt.t

get_liquidity_balance c env state returns the amount of liquidity token owned by c in the concrete part of state, according to the Liquidity contract.

val get_cpmm_total_liquidity : Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t env -> t -> liquidity Tezos_base.TzPervasives.tzresult Lwt.t

get_cpmm_total_liquidity env state fetches the current amount of liquidity tokens distributed by the CPMM contract using the concrete part of the state state.