Module Slashing_helpers.Full_denunciation

Helpers about "full denunciations", that is, a denunciation item and its culprit. See type t of this module.

type t = Tezos_base.TzPervasives.Signature.Public_key_hash.t * Tezos_protocol_020_PsParisC.Protocol.Denunciations_repr.item

A denunciation item preceded by the culprit's pkh. Indeed, the culprit isn't recorded inside the Protocol.Denunciations_repr.item because it serves as a key in the protocol's storage instead. But we often need both together in the tests.

val check_same_lists_any_order : loc:string -> t list -> t list -> unit Tezos_base.TzPervasives.tzresult Lwt.t

Asserts that both lists contain the same elements.

These elements may be ordered differently, but must have the same multiplicity in both lists.