Module Tezos_020_PsParisC_test_helpers.Transfers

val transfer_and_check_balances : ?with_burn:bool -> loc:string -> Incremental.t -> ?fee:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> ?expect_apply_failure: (Tezos_base.TzPervasives.error Tezos_base.TzPervasives.trace -> unit Tezos_base.TzPervasives.tzresult Lwt.t) -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> (Incremental.t * Tezos_protocol_020_PsParisC.Protocol.Alpha_context.packed_operation) Tezos_base.TzPervasives.tzresult Lwt.t

transfer_and_check_balances b fee src dst amount this function takes a block, an optional parameter fee if fee does not given it will be set to zero tez, a source contract, a destination contract and the amount that one wants to transfer.

1- Transfer the amount of tez (w/wo fee) from a source contract to a destination contract.

2- Check the equivalent of the balance of the source/destination contract before and after transfer is validated.

This function returns a pair:

A block that added a valid operation

a valid operation

val n_transactions : int -> Incremental.t -> ?fee:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Contract.t -> Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Tez.t -> Incremental.t Tezos_base.TzPervasives.tzresult Lwt.t

n_transactions n b fee source dest amount this function takes a number of "n" that one wish to transfer, a block, an optional parameter fee, a source contract, a destination contract and an amount one wants to transfer.

This function will do a transaction from a source contract to a destination contract with the amount "n" times.