Module Baking_state.Delegate_slots

type t

Information regarding the slot distribution at some level.

val own_delegates : t -> delegate_slot list

Returns the list of our own delegates that have at least a slot. There are no duplicates, the associated slot is the first one.

val own_slot_owner : t -> slot:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Slot.t -> delegate_slot option

Returns, among our *own* delegates, the delegate (together with its first attesting slot) that owns the given slot, if any (even if the given slot is not the delegate's first slot).

val voting_power : t -> slot:Tezos_protocol_020_PsParisC.Protocol.Alpha_context.Slot.t -> int option

Returns the voting power of the delegate whose first slot is the given slot. Returns None if the slot is not the first slot of any delegate.