Module Tezos_client_020_PsParisC

module Annotated_manager_operation : sig ... end

Annotated manager operations are wrappers used to accumulate information (especially about limits) on the operation prior to the injection.

module Client_proto_args : sig ... end

module Client_proto_context : sig ... end

module Client_proto_contracts : sig ... end

module Client_proto_fa12 : sig ... end

Client_proto_fa12 implements built-in support for the FA1.2 standard. This module features functions to check whether a contract implements the standard interface and to interact with such contracts using high-level actions that model the entrypoint calls.

module Client_proto_multisig : sig ... end

module Client_proto_programs : sig ... end

module Client_proto_rollups : sig ... end

module Client_proto_tzt : sig ... end

module Client_proto_utils : sig ... end

module Injection : sig ... end

module Light : sig ... end

module Limit : sig ... end

This helper module allows to handle partially specified limits during the injection process.

module Managed_contract : sig ... end

module Michelson_v1_emacs : sig ... end

module Michelson_v1_entrypoints : sig ... end

module Michelson_v1_error_reporter : sig ... end

module Michelson_v1_helpers : sig ... end

module Michelson_v1_macros : sig ... end

module Michelson_v1_parser : sig ... end

module Michelson_v1_printer : sig ... end

module Michelson_v1_stack : sig ... end

module Mockup : sig ... end

module Operation_result : sig ... end

module Protocol_client_context : sig ... end

module Proxy : sig ... end