Class Protocol_client_context.wrap_full

From a Client_context.full, the class allows to call RPCs from the node and those defined by the protocol.

inherit Tezos_client_base.Client_context.full

The class Client_context.full provides I/O services for the client, the wallet, etc.

inherit (Tezos_shell_services.Shell_services.chain * Tezos_shell_services.Shell_services.block) Tezos_protocol_020_PsParisC.Environment.RPC_context.simple

Base interface provided to call RPCs, i.e., communication with the node. A client context is defined by mapping all RPCs protocol-generic to a specific protocol.

inherit [Tezos_shell_services.Shell_services.chain, Tezos_shell_services.Shell_services.block] Tezos_protocol_020_PsParisC.Environment.proto_rpc_context

Protocol RPCs exposed through the environment (using an additional chainpath).